updated both, but it's still untested.
here's what I did.
- I moved the creation of the rental object client-side. Now, when the
user clicks the button, the function "other" will create a rental object
and set the column "paid" as false.
- I updated the serverside code to correspond to this change, removing
the creation of the rental object and just updating it, as we have been
able to do before.
- I moved the chuck of code inside of the Parse.User.become block
because the code should not be allowed to run if the user did not
complete his payment successfully. I added a failsafe to ensure that if
payment failed, the code would not be executed and everything stays as
is.